Extract core lib #33
