+++ apps/calc-web/src/app/components/base-converter/base-converter-component.tsx
@@ -3,7 +3,6 @@ import { BaseDigits, ComplementConverter, Conversion, fromString, isValidString 
 import { SwapOutlined } from '@ant-design/icons/lib';
 import { InputWithCopy } from '@calc/ui';
 import { useSelector } from 'react-redux';
-import { selectShowComplement, selectShowDecimalValue } from '../../store/selectors/options.selectors';
 import { ConversionOptions } from '../conversion-options/conversion-options';
 import { useTranslation } from 'react-i18next';
 import { Button, IconButton, TextField, Tooltip } from '@material-ui/core';
@@ -11,6 +10,7 @@ import { clean } from '@calc/utils';
 import { useFormik } from 'formik';
 import { FormErrors } from '@calc/ui';
 import { useConverterStyles } from '../../core/styles/converter-styles';
+import { selectShowComplement, selectShowDecimalValue } from '@calc/core';
 
 interface P {
     onConversionChange?: (conversion: Conversion, precision: number) => void;
