+++ apps/calc-web/src/app/components/conversion-options/conversion-options.tsx
@@ -1,10 +1,9 @@
 import React, { CSSProperties, FC } from 'react';
-import { setShowComplement, setShowDecimalValue } from '../../store/actions/options.actions';
 import { useDispatch, useSelector } from 'react-redux';
-import { selectShowComplement, selectShowDecimalValue } from '../../store/selectors/options.selectors';
 import { useTranslation } from 'react-i18next';
 import { createStyles, FormControlLabel, FormGroup, Switch, Theme } from '@material-ui/core';
 import { makeStyles } from '@material-ui/core/styles';
+import { selectShowComplement, selectShowDecimalValue, setShowComplement, setShowDecimalValue } from '@calc/core';
 
 interface P {
     style?: CSSProperties;
