+++ apps/calc-web/src/app/components/language-menu/language-menu.tsx
@@ -3,9 +3,8 @@ import { useTranslation } from 'react-i18next';
 import { Button, Popover } from '@material-ui/core';
 import TranslateIcon from '@material-ui/icons/Translate';
 import { useDispatch, useSelector } from 'react-redux';
-import { selectAppLanguage } from '../../store/selectors/options.selectors';
-import { setLanguage } from '../../store/actions/options.actions';
 import { availableLanguages, getNativeName, Language } from '@calc/i18n';
+import { selectAppLanguage, setLanguage } from '@calc/core';
 
 export const LanguageMenu: FC = () => {
     const { i18n, t } = useTranslation();
