+++ apps/calc-web/src/main.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './app/app';
 import { Provider } from 'react-redux';
-import { store } from './app/store/configure-store';
+import { store } from '@calc/core';
 
 ReactDOM.render(
     <Provider store={store}>
