+++ libs/core/README.md
@@ -1,0 +1,7 @@
+# core
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test core` to execute the unit tests via [Jest](https://jestjs.io).
