+++ libs/core/jest.config.js
@@ -1,0 +1,9 @@
+module.exports = {
+  name: 'core',
+  preset: '../../jest.config.js',
+  transform: {
+    '^.+\\.[tj]sx?$': 'ts-jest',
+  },
+  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'html'],
+  coverageDirectory: '../../coverage/libs/core',
+};
