+++ libs/core/src/index.ts
@@ -1,0 +1,4 @@
+export * from './lib/store/configure-store'
+export * from './lib/store/actions/options.actions'
+export * from './lib/store/selectors/options.selectors'
+
