+++ libs/core/src/lib/store/actions/options.actions.ts
@@ -1,8 +1,8 @@
 import { createAction } from '@reduxjs/toolkit';
 import { AppTheme } from '@calc/ui';
-import { loadOptionsWithFallback, updateOptions } from '../../core/functions/local-storage';
 import { getI18n } from 'react-i18next';
 import { Language } from '@calc/i18n';
+import { loadOptionsWithFallback, updateOptions } from '../../functions/local-storage';
 
 export enum OptionsActionsType {
     SetShowComplement = '[Options] Set show complement',
