+++ libs/core/src/lib/store/configure-store.ts
