+++ libs/core/src/lib/store/selectors/options.selectors.ts
