+++ nx.json
@@ -50,6 +50,9 @@
     },
     "i18n": {
       "tags": []
+    },
+    "core": {
+      "tags": []
     }
   }
 }
