+++ tsconfig.json
@@ -28,7 +28,8 @@
       "@calc/positional-calculator": [
         "libs/positional-calculator/src/index.ts"
       ],
-      "@calc/i18n": ["libs/i18n/src/index.ts"]
+      "@calc/i18n": ["libs/i18n/src/index.ts"],
+      "@calc/core": ["libs/core/src/index.ts"]
     }
   },
   "exclude": ["node_modules", "tmp"]
